requirements.txt
docopt==0.6.2
ImageHash==3.4
more-itertools==2.2
Pillow==4.3.0
python-magic==0.4.15
tinydb===3.15.1